+++ Libraries/Jquery.php
@@ -106,6 +106,9 @@ class Jquery
                 break;
         }
 
+        //Removes defer attribute
+        $data['defer'] = false;
+
         return $data;
     }
 }