+++ Libraries/Jqueryslim.php
@@ -95,6 +95,9 @@ class Jqueryslim
                 break;
         }
 
+        //Removes defer attribute
+        $data['defer'] = false;
+
         return $data;
     }
 }